Refactor Printer and enhance interpreter functionality
This commit introduces significant changes to the codebase, including:
- Refactoring the `Printer` class into `PrinterExprVisitor` and `PrinterStmtVisitor` for improved modularity.
- Adding the `ClockCallable` class to provide a callable function for retrieving the current time.
- Updating the `INoxCallable` interface and enhancing the `Interpreter` class to support a global environment and time retrieval.
- Expanding the `Keywords` class with additional keywords for classes, functions, and control flow.
- Refactoring the `Parser` and `Resolver` classes to improve parsing logic and variable/function scope resolution.
- Updating the `Tokenizer` to recognize the new keywords.

These changes enhance the structure, readability, and overall functionality of the interpreter.

+++ Nox/Callables/ClockCallable.cs
@@ -1,4 +1,5 @@
 ﻿
+using Nox.InterpreterImpl;
 
 namespace Nox.Callables
 {
@@ -11,7 +12,7 @@ namespace Nox.Callables
 
         public object Call(Interpreter interpreter, List<object> arguments)
         {
-            return (double)DateTimeOffset.UtcNow.ToUnixTimeMilliseconds() / 1000.0;
+            return DateTimeOffset.UtcNow.ToUnixTimeMilliseconds() / 1000.0;
         }
 
         public override string ToString() => "<native fn>";
